mf4 stream: an exact block read in the loader, a probed skip, a DZ cap the format states, 64-bit ordinals (codex #342 round 4)
Four findings, all true:

- The loader's block read went through `bytes_at`, which ZERO-FILLS what it could not read (P2):
  since load_file reads through a FileSource now, a truncation or a failed read handed the
  decoder fabricated records and called them a recording. `block_bytes` reads exactly and fails
  short, as `fill` does.
- A skipped plain span was stepped over by offset alone (P2), so a file truncated under the
  reader inside a skipped record read as a clean end. The span's LAST byte is probed — the walker
  clamped every block to the file at open, so a truncation ends inside a span and the probe fails
  on it. Stated limit: a device failing a read mid-span whose end it serves; reading and
  discarding would see it, at the cost of reading every byte of a group this reader does not
  decode, which is the loader's cost and the one the stream exists not to pay.
- `max_dz_block` was 256 MiB, and it is the memory PER DATA GROUP (P2): every sorted cursor may
  hold one inflated block (and one in its signal-data view) while its head waits in the merge,
  so a valid multi-bus file was gigabytes before its first row. 16 MiB now — the format says an
  uncompressed data block should not exceed 4 MB, and this is four times that — and a block is
  released the moment it is served whole, so a cursor waiting between blocks holds none.
- Record ordinals were int (P2): a recording of tens of GB has more records than an int counts,
  and a wrapped ordinal reorders equal-timestamp rows across groups. u64 in the cursor and both
  stored positions.

Tests: a short read inside a plain block fails the loader; a skip over a span the source
truncates fails the stream (`short read`) and the loader; a DZ block served whole is released.

+++ modules/mf4/stream.v
@@ -307,16 +307,18 @@ fn (u &UCg) exhausted() bool {
 	return !u.ok || (u.seen >= u.cap && u.deferred.len == 0)
 }
 
-// RawRec is a record held back, and its position.
+// RawRec is a record held back, and its position. Positions are u64: a recording of tens of
+// GB has more records than an int counts, and a wrapped ordinal reorders equal-timestamp rows
+// (codex on #342 round 4).
 struct RawRec {
 	raw []u8
-	pos int
+	pos u64
 }
 
 // Pending is a decoded row and the record position it came from.
 struct Pending {
 	row canlog.Row
-	pos int
+	pos u64
 }
 
 // The read-ahead cap over an unsorted group's queues, in rows: eight thousand rows a queue of
@@ -349,7 +351,7 @@ mut:
 	views       map[int]&ChainView
 	stream      ChainStream
 	rec_id_size int
-	rec_n       int // records read so far: the ordinal the loader keys ties by
+	rec_n       u64 // records read so far: the ordinal the loader keys ties by
 	queued      int
 	deferred    int // frames held back across all groups
 	deferred_b  u64 // and their bytes
@@ -456,7 +458,7 @@ fn payload_ready(lay &CgLayout, raw []u8, base int, ring &RingVlsd) bool {
 
 // decode_into decodes one fixed record of group `ci` into its queue, through whatever its
 // payload source is.
-fn (mut c UnsortedCursor) decode_into(ci int, raw []u8, base int, pos int, mut log canlog.Log) {
+fn (mut c UnsortedCursor) decode_into(ci int, raw []u8, base int, pos u64, mut log canlog.Log) {
 	mut u := &c.cgs[ci]
 	row := if mut ring := c.vlsd[u.lay.vlsd_link] {
 		decode_row(&u.lay, raw, base, mut ring, mut u.labels, mut log)
